Fix TrialShortcut fast-track having no efficacy penalty at disease gen 0
The fast-track option set strain calibration to current_gen.saturating_sub(2),
which at gen 0 produced 0 — zero drift, zero penalty. This made fast-track
strictly better than maintaining standards for any disease that hadn't
mutated yet (the common case for most pathogen types).

Changed strain_generations from Vec<u32> to Vec<i32> so fast-tracking at
gen 0 produces gen -2, giving the intended 30% efficacy penalty regardless
of disease generation.

Closes #682

+++ src/engine/research.rs
@@ -119,7 +119,7 @@ pub(super) fn tick_research(state: &mut GameState) {
                         // Update strain calibration to current disease generation
                         if let Some(pos) = medicine.target_diseases.iter().position(|&d| d == d_idx) {
                             let current_gen = state.diseases.get(d_idx)
-                                .map_or(0, |d| d.strain_generation);
+                                .map_or(0, |d| d.strain_generation) as i32;
                             // Extend strain_generations if needed
                             while medicine.strain_generations.len() <= pos {
                                 medicine.strain_generations.push(0);
@@ -154,7 +154,7 @@ pub(super) fn tick_research(state: &mut GameState) {
                         // Calibrate to current strain generations of all target diseases
                         medicine.strain_generations = medicine.target_diseases.iter()
                             .map(|&d_idx| state.diseases.get(d_idx)
-                                .map_or(0, |d| d.strain_generation))
+                                .map_or(0, |d| d.strain_generation as i32))
                             .collect();
                     }
                 }
